pyproject.toml
[tool.poetry]
name = "ocr-receipt"
version = "0.1.0"
description = "A desktop application to extract structured data from PDF invoices using OCR."
authors = ["Martin Fournier <martinfou@gmail.com>"]

[[tool.poetry.packages]]
include = "ocr_receipt"
from = "src"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
pyqt6 = "^6.4.0"
pytesseract = "^0.3.10"
pdf2image = "^1.16.0"
opencv-python = "^4.7.0"
pillow = "^11.3.0"
pypdf2 = "^3.0.0"
click = "^8.0.0"
pyyaml = "^6.0"
yoyo-migrations = "^9.0.0"
python-dateutil = "^2.9.0.post0"
pypdf = "^5.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-qt = "^4.0.0"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"
yoyo-migrations = "^9.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
